pyproject.toml
[project]
name = "trinity-turbo"
version = "0.1.0"
description = "Layer-aware KV cache compression for Trinity-Large-Thinking on vLLM — optimized for SM120 Blackwell"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Lna-Lab", email = "lna-lab@users.noreply.github.com" },
]
readme = "README.md"
keywords = ["vllm", "kv-cache", "quantization", "moe", "blackwell", "sm120", "trinity"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "vllm>=0.19.0",
    "torch>=2.6.0",
    "triton>=3.2.0",
]

[project.optional-dependencies]
cutile = ["cuda-tile[tileiras]"]
dev = [
    "pytest>=8.0",
    "pytest-benchmark",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trinity_turbo"]

[project.entry-points."vllm.general_plugins"]
trinity_turbo = "trinity_turbo.plugin:register_trinity_turbo"

[tool.pytest.ini_options]
testpaths = ["tests"]
